+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## master
 
+* Adds the zone data-source
 * Adds the domain resource (to create domains in DNSimple)
 
 ## 0.6.0